+++ Assets/Editor/Resource.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b14d44e5f6e7ad458299af58dae16db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
