+++ Assets/Scripts/Interactables/Resource/Resource.cs
@@ -1,55 +1,59 @@
 using Examen.Inventory;
 using Examen.Items;
+using Examen.Poolsystem;
 using MarkUlrich.Health;
 using System;
 using System.Collections;
 using UnityEngine;
 
-[RequireComponent(typeof(HealthData))]
-public class Resource : MonoBehaviour, Interactable
+namespace Examen.Interactables.Resource
 {
-    [HideInInspector]public HealthData HealthData;
-    [HideInInspector]public PoolSystem poolSystem;
-
-    public Item ResourceItem;
-    public int AmountToGive = 1;
-    public int DeathTime;
-
-    public const int DamageAmount = 1;
-    public Action OnRespawn;
-
-    public virtual void Interact()
+    [RequireComponent(typeof(HealthData))]
+    public class Resource : MonoBehaviour, Interactable
     {
-        Debug.Log("interact");
-        PlayInteractingSound();
-        //playanimation
-        HealthData.TakeDamage(DamageAmount);
-        InventorySystem.AddItem(ResourceItem, AmountToGive);
-    }
+        [HideInInspector] public HealthData HealthData;
+        [HideInInspector] public PoolSystem poolSystem;
 
-    public virtual void PlayInteractingSound()
-    {
+        public Item ResourceItem;
+        public int AmountToGive = 1;
+        public int DeathTime;
 
-    }
+        public const int DamageAmount = 1;
+        public Action OnRespawn;
 
-    public virtual void Start()
-    {
-        poolSystem = PoolSystem.Instance;
-        HealthData = GetComponent<HealthData>();
-        HealthData.onDie.AddListener(StartDeathTimer);
-    }
+        public virtual void Interact()
+        {
+            Debug.Log("interact");
+            PlayInteractingSound();
+            //playanimation
+            HealthData.TakeDamage(DamageAmount);
+            InventorySystem.AddItem(ResourceItem, AmountToGive);
+        }
 
-    public virtual void StartDeathTimer() 
-    {
-        poolSystem.DespawnObject(ResourceItem.Name);
-        StartCoroutine(DeathTimer());
-    }
+        public virtual void PlayInteractingSound()
+        {
 
-    private IEnumerator DeathTimer()
-    {
-        yield return new WaitForSeconds(DeathTime);
+        }
 
-        poolSystem.SpawnObject(ResourceItem.Name);
-        OnRespawn?.Invoke();
+        public virtual void Start()
+        {
+            poolSystem = PoolSystem.Instance;
+            HealthData = GetComponent<HealthData>();
+            HealthData.onDie.AddListener(StartDeathTimer);
+        }
+
+        public virtual void StartDeathTimer()
+        {
+            poolSystem.DespawnObject(ResourceItem.Name);
+            StartCoroutine(DeathTimer());
+        }
+
+        private IEnumerator DeathTimer()
+        {
+            yield return new WaitForSeconds(DeathTime);
+
+            poolSystem.SpawnObject(ResourceItem.Name);
+            OnRespawn?.Invoke();
+        }
     }
 }
