--- Assets/Scripts/Interactables/Resource/Stone.cs
@@ -1,5 +1,0 @@
-using Examen.Inventory;
-public class Stone : Resource
-{
-
-}
