--- Assets/Scripts/Interactables/Resource/Stone.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 51666943ec35f2748b9ac69331c59334
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
