--- Assets/Scripts/Interactables/Resource/Tree.cs
@@ -1,5 +1,0 @@
-public class Tree : Resource
-{
-
-
-}
