--- Assets/Scripts/Interactables/Resource/Tree.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: efd769099e2bb664e81be915184e3cc3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
